+++ experiment3/transformer-vae/data_processing.py
@@ -2,6 +2,22 @@ import json
 import numpy as np
 import pandas as pd
 from sklearn.model_selection import train_test_split
+from torch.utils.data import Dataset
+
+class CustomDataset(Dataset):
+    def __init__(self, texts, rejected, tokenizer, max_length):
+        self.texts = texts
+        self.rejected = rejected
+        self.tokenizer = tokenizer
+        self.max_length = max_length
+
+    def __len__(self):
+        return len(self.texts)
+
+    def __getitem__(self, idx):
+        text = self.texts[idx]
+        inputs = self.tokenizer(text, padding='max_length', truncation=True, max_length=self.max_length, return_tensors="pt")
+        return inputs.input_ids.squeeze(0), inputs.attention_mask.squeeze(0), self.rejected[idx]
 
 # Define how to load and preprocess the data.
 def preprocess_data(file_path, text_source):
